requirements.txt
streamlit==0.89.0
tensorflow==2.6.0
numpy==1.21.2
opencv-python-headless==4.5.3.56
pillow==8.3.2